backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic[email]==2.10.4
pydantic-settings==2.7.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
redis[asyncio]==5.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
docker==7.1.0
huggingface_hub==0.27.0
structlog==24.4.0
pynvml==11.5.3
python-multipart==0.0.20
